requirements.txt
yt-dlp
openai-whisper

# System dependency (install separately):
# ffmpeg (must be installed and added to PATH)